feat: move users related commands to own module
