+++ app/tests/modules/aws/test_aws.py
@@ -34,24 +34,24 @@ def test_aws_command_handles_access_command(request_access_modal):
     request_access_modal.assert_called_with(client, body)
 
 
-@patch("modules.aws.aws.slack_commands.parse_command")
-@patch("modules.aws.aws.request_user_provisioning")
-def test_aws_command_handles_provision_command(
-    mock_request_provisioning: MagicMock, mock_parse_command: MagicMock
-):
-    ack = MagicMock()
-    logger = MagicMock()
-    respond = MagicMock()
-    client = MagicMock()
-    body = MagicMock()
-    mock_parse_command.return_value = ["user", "create", "user.name@email.com"]
-    aws.aws_command(
-        ack, {"text": "user create user.name@email.com"}, logger, respond, client, body
-    )
-    ack.assert_called()
-    mock_request_provisioning.assert_called_with(
-        client, body, respond, ["create", "user.name@email.com"], logger
-    )
+# @patch("modules.aws.aws.slack_commands.parse_command")
+# @patch("modules.aws.aws.request_user_provisioning")
+# def test_aws_command_handles_provision_command(
+#     mock_request_provisioning: MagicMock, mock_parse_command: MagicMock
+# ):
+#     ack = MagicMock()
+#     logger = MagicMock()
+#     respond = MagicMock()
+#     client = MagicMock()
+#     body = MagicMock()
+#     mock_parse_command.return_value = ["user", "create", "user.name@email.com"]
+#     aws.aws_command(
+#         ack, {"text": "user create user.name@email.com"}, logger, respond, client, body
+#     )
+#     ack.assert_called()
+#     mock_request_provisioning.assert_called_with(
+#         client, body, respond, ["create", "user.name@email.com"], logger
+#     )
 
 
 @patch("modules.aws.aws.slack_commands.parse_command")
@@ -96,58 +96,6 @@ def test_aws_command_handles_unknown_command(mock_parse_command):
     ack.assert_called()
     respond.assert_called_with(
         "Unknown command: `unknown`. Type `/aws help` to see a list of commands.\nCommande inconnue: `unknown`. Tapez `/aws help` pour voir une liste des commandes."
-    )
-
-
-@patch("modules.aws.aws.AWS_ADMIN_GROUPS", ["admin-group@email.com"])
-@patch("modules.aws.aws.provision_aws_users")
-@patch("modules.aws.aws.permissions")
-@patch("modules.aws.aws.slack_users")
-def test_request_user_provisioning(
-    mock_slack_users, mock_permissions, mock_provision_aws_user
-):
-    client = MagicMock()
-    body = MagicMock()
-    respond = MagicMock()
-    logger = MagicMock()
-    mock_slack_users.get_user_email_from_body.return_value = "user.name@email.com"
-    mock_provision_aws_user.return_value = True
-    aws.request_user_provisioning(
-        client, body, respond, ["create", "user.email"], logger
-    )
-    mock_slack_users.get_user_email_from_body.assert_called_with(client, body)
-    mock_permissions.is_user_member_of_groups.assert_called_with(
-        "user.name@email.com", ["admin-group@email.com"]
-    )
-    respond.assert_called_with("Request completed:\ntrue")
-    logger.info.assert_called_with("Completed user provisioning request")
-
-
-@patch("modules.aws.aws.AWS_ADMIN_GROUPS", ["admin-group@email.com"])
-@patch("modules.aws.aws.provision_aws_users")
-@patch("modules.aws.aws.permissions")
-@patch("modules.aws.aws.slack_users")
-def test_request_user_provisioning_requestor_not_admin(
-    mock_slack_users,
-    mock_permissions,
-    mock_provision_aws_user,
-):
-    client = MagicMock()
-    body = MagicMock()
-    respond = MagicMock()
-    logger = MagicMock()
-    mock_slack_users.get_user_email_from_body.return_value = "notadmin.name@email.com"
-    mock_permissions.is_user_member_of_groups.return_value = False
-    mock_provision_aws_user.return_value = True
-    aws.request_user_provisioning(
-        client, body, respond, ["create", "user.email"], logger
-    )
-    mock_slack_users.get_user_email_from_body.assert_called_with(client, body)
-    mock_permissions.is_user_member_of_groups.assert_called_with(
-        "notadmin.name@email.com", ["admin-group@email.com"]
-    )
-    respond.assert_called_with(
-        "This function is restricted to admins only. Please contact #sre-and-tech-ops for assistance."
     )
